requirements.txt
telethon
python-dotenv
pandas
plotly
tqdm
schedule==1.2.1
